Replace render.sh with a more comprehensive glimpse-render.py
The render.sh was just a quick script for kicking off multiple instances
of Blender to render training data but was overly crude.

This updated front-end script has a clearer interface documented via
--help and supports a --dry-run option that can conveniently tell you how
many frames a particular run would result in which helps with turning
a run to result in a certain size data set.