Add joint map from glimpse/src/joint-map.json
This seems like a more appropriate location for the joint map which is
used as an input for training.

The joint map has been added as joint-maps/2018-06-joint-map.json
where the prefix should match the -to-YEAR-MONTH-rdt-map.json suffix
of the label map given to the pre-processor.

The expectation is that further joint maps will be added for use with
alternative label maps.