generator: add camera, fixed clothes, background and tags features
- add fixed/debug/smooth camera features in the parameters to have
better control of how the camera is being positioned during rendering
process of the images

- add background build feature that allows us to progressively build a
wall and floor to include in the render. This feature allows us
to render an artificial video resembling real world recording for
glimpse viewer testing.

- add fixed clothes feature that allows us to specify a specific set of
clothes we want all models to wear in the rendered images.

- add `--tags-whitelist`, `--tags-blacklist` and `--tags-skip`
parameters in order to better control which bvh files we want to
exclude/include in the rendering process basing on the tags specified
in the mocap index file.