pre-processor-configs: describe foreground-edge-swizzle noise
The pre-processor has generalised how it can add noise to the silhouette
of the body with a new 'foreground-edge-swizzle' noise type. This
updates the configs to use this (matching the implicit behaviour of the
pre-processor before this was generalised). This also adds a small
blurb about the this in pre-processor-configs/README.md