glimpse_data_generator: Formalize+generalize index filtering
Initially filtering which bvh mocap sequences we want to process was
done with a [start:end) range based on the indices for sequences in our
index.json file. More recently we've added support for filtering with
white and black lists.

It's likely we will want more ways of filtering (such as matching the
name of particular sequences), or only matching ones that have been
preloaded.

There are multiple places where we want filtering but broadly speaking
there are two notably different filtering use cases:
1) When reviewing and navigating sequences via the UI - where the filter
   parameters are controlled interactively.
2) For each of our non-interactive operations (info, preload, purge and
   render) - where the filter parameters are setup by
   glimpse-generator.py

Notably there was some duplication and inconsistency with the logic
handling filtering for these different cases so this change introduces
two new shared classes that help formalize how we open and filter an
index:

The BvhIndex handles parsing a full index.json and normalizing entries
to simplify querying each entry. The BvhFilteredIndex is given a full
BvhIndex and then a set of filter options (including the start:end range
and white/black lists) to create a reduced index.  Both the full or
filtered indexes can be accessed as an iterator.

A load_filtered_mocap_index() utility handles loading a full index and
then creating a filtered index for all the non-interactive operations we
have.

A update_ui_filtered_index() utility is called via the UI code whenever
some property affecting filtering is changed and this will build a new
BvhFilteredIndex from the global BvhIndex.

Previously everything was building on top of the globally opened
bvh_index but to help reduce the amount of global state we share across
very different use cases we now only maintain a global index for the UI
code, renamed to ui_full_bvh_index_obj. There is also a
ui_filtered_bvh_index_obj that's updated via update_ui_filtered_index().
All of the non-interactive operations are now more orthogonal since they
load and filter the index on-the-fly, locally when the operator is
invoked.

Less emphasis is put on reporting the [start:end) range considering
that these indices aren't really very intuitive and we expect to add
alternative ways of filtering and controlling the maximum number of
sequences we iterate.