glimpse-generator: support .json render configs
This was done for a few reasons:
1) The command line interface for glimpse-generator was getting
   unwieldy.
2) It's more awkward to review how older data sets were created
   if rendering is only configured via command line switches.

Via a .json config, specified with --config=, we can now configure
filtering details (which tags to white/black list), camera details
(resolution and position) and the rate that clothing should be
randomized or whether a background should be included.

This adds a render-configs/iphone-x-training.json as a first config
and a render-configs/README.md with an overview of the schema.

Since the schema design wasn't directly based on the command line
argument interface some addon changes were made to accommodate
a slightly different configuration model. In particular the fixed
body and clothing options are now configured as body/clothing
whitelists which are orthogonal from the positioning mode of the
camera.