glimpse_data_generator: update collection and reports of tag stats
This adjusts how we collect stats about tags in relation to frames that
are skipped or rendered.

Previously if we skipped a frame due to a tag skip percentage then we
would increment a single tag counter for the tag that happened to
trigger the skip first. This didn't consider that it's possible that
multiple tags could be associated with the skipped bvh that would have
also resulted in skipping. Really though, the _reason_ for skipping
isn't really that important to us in the end, compared to, say, having
statistics that summarise the tags of all frames skipped.

This updates how we collect stats related to [tag] skip percentages so
we update counts for all tags associated with a skipped bvh sequence
(instead of the counter for the tag causing a skip).

Even though we now collect stats for frames skipped due to [tag] skip
percentages we're notably less interested in stats for frames skipped
for one particular reason ([tag] skip percentages) than in stats for
_all_ skipped frames, so for now we actually stop reporting any stats
related to skipped frames. Maybe later we can combine our filtering
stats with skip percentage stats for a more complete overview of skipped
frames.

Overall our main interest is in statistics of frames that we _do_
render, not frames that we skip so this adds those stats to the final
report.