glimpse-build-training-jobs: support templates
This adds support for a job template .json file that can provide default
values for properties and support for index and out_file template
strings.

Within an index or out_file template (--index-template /
--out-file-template) then '{job}' expands to the unique index of the
job. Within an out_file template then '{index}' expands to the value of
of the 'index' name property.

This also makes some pedantic flake8 warning fixes; adds some comments
for things I was initially a bit unsure of and re-works the recursion
to simplify a bit (avoiding a special case for the first job).