glimpse-build-training-jobs: remove special index/out_file template args
Instead of having --index-template and --out-file-template options we
just handle expanding {} variables in any given values for 'out_file' or
'index'.

This adds a {date} variable that will expand to a YYYY-MM-DD value.

It would be nice to generalise this farther and allow arbitrary
cross-referencing of other properties but that would involve sorting
the precedence of properties based on variable dependencies and the
added complexity isn't currently justified here.