glimpse-mocap-stats: give stats for indexed frames not full mocaps/index.json
This updates glimpse-mocap-stats to load an index that was created via
glimpse-data-indexer.py and report the distribution of tags across those
indexed frames.

This removes all the filtering options (e.g. tag white/blacklists etc)
since glimpse-data-indexer can handle filtering.

Stats for the full mocaps/index.json can be more readily got via
glimpse-generator render --dry-run which will show the distribution of
tags across the frames that would be rendered.

We were missing a way to get a quick overview of the tags across the
frames in an index created with glimpse-data-indexer.