glimpse-generator: set GlimpseBvhTagsSkip='none' for empty tag_skip_percentages
The default value for GlimpseBvhTagsSkip is 'none' so it wasn't
consistent that an empty tag_skip_percentages filter would result in an
empty '' value instead of 'none'.

Just in case we later introduce some other case where we leave
GlimpseBvhTagSkip unset we will continue to check for it being unset in
the addon, but this also re-adds a check for 'none' that was recently
removed in commit 431402c99f68ea9499.